refactor: 컨테이너로 생성 로직 위임 및 패키지 이동
