+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/HandlerAdapterRegistry.java
@@ -1,16 +1,16 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet.mvc.tobe;
 
 import java.util.List;
 
-import com.interface21.webmvc.servlet.mvc.tobe.AnnotationHandlerAdapter;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerAdapter;
+import com.interface21.context.container.Container;
 
 public class HandlerAdapterRegistry {
 
     private final List<HandlerAdapter> handlerAdapters;
 
     public HandlerAdapterRegistry() {
-        handlerAdapters = List.of(new ManualHandlerAdaptor(), new AnnotationHandlerAdapter());
+        Container container = Container.getInstance();
+        handlerAdapters = container.getInstancesOf(HandlerAdapter.class);
     }
 
     public HandlerAdapter getHandlerAdapter(final Object handler) {
