+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/HandlerMappingRegistry.java
@@ -1,19 +1,19 @@
-package com.techcourse;
+package com.interface21.webmvc.servlet.mvc.tobe;
 
 import java.util.List;
 import java.util.stream.Stream;
 
 import jakarta.servlet.http.HttpServletRequest;
 
-import com.interface21.webmvc.servlet.mvc.tobe.AnnotationHandlerMapping;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerMapping;
+import com.interface21.context.container.Container;
 
 public class HandlerMappingRegistry {
 
     private final List<HandlerMapping> handlerMappings;
 
-    public HandlerMappingRegistry(final Object... basePackages) {
-        handlerMappings = List.of(new ManualHandlerMapping(), new AnnotationHandlerMapping(basePackages));
+    public HandlerMappingRegistry() {
+        Container container = Container.getInstance();
+        handlerMappings = container.getInstancesOf(HandlerMapping.class);
         handlerMappings.forEach(HandlerMapping::initialize);
     }
